[ProtonAOSP][rvc] SystemUI: Remove divider between QS tiles and media player
The color used for this divider in dark mode contrasts too much with the
pure black background in the QS, and even in light mode, the divider
disrupts the otherwise clean and streamlined style of not having
dividers in the QS. Let's just remove the divider altogether for both
themes to clean up the QS when the media player is shown.

The code controlling the divider has been removed altogether to avoid
misleading people who read through the code, as well as to reduce unused
code bloat.

Change-Id: Ie3d9e0070c974b598a8784225375882b4b67173e